requirements-train.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
Pillow==10.4.0
joblib==1.4.2

